lc3sim_project

Module defs

Structs

InvalidRegAddr

Enums

Op

OpDiscriminants

Auto-generated discriminant enum variants

PseudoOp

PseudoOpDiscriminants

Auto-generated discriminant enum variants

RegAddr

Constants

ADDR_SPACE_SIZE

Size of the memory address space.

DEV_REG_ADDR

First address of the device register address space.

DISPLAY_DATA_REGISTER

Characters written to this low byte will be displayed on screen.

DISPLAY_STATUS_REGISTER

Bit 15 is 1 when the display is ready to receive a new character.

IO_PRIORITY

Priority for an I/O interrupt

IR_VEC_TBL

First address of the interrupt vector table.

KEYBOARD_DATA_REGISTER

Last character typed on the keyboard.

KEYBOARD_INTERRUPT

Vector for a keyboard I/O interrupt.

KEYBOARD_STATUS_REGISTER

Bit 15 is 1 when the keyboard has received a new character.

MACHINE_CONTROL_REGISTER

Bit 15 is clock enable – machine runs while set to 1.

NUM_REGS

Number of registers in the LC3 spec

OS_SUPER_STACK

First address of the operating and supervisor stack space.

STACK_REG

Register with a special stack meaning in privileged mode.

SUPERVISOR_SP_INIT

Initial supervisor stack pointer value.

TRAP_VEC_TBL

First address of the trap vector table.

USER_SPACE

First address of the user code space.

Type Aliases

LC3MemAddr

LC3Word

SignedLC3Word