lc3sim_project::executors
Trait LC3

pub trait LC3 {
    type FullIter<'a>: Iterator<Item = LC3Word>
       where Self: 'a;
    type SparseIter<'a>: Iterator<Item = LC3MemLoc>
       where Self: 'a;

    // Required methods
    fn pc(&self) -> LC3MemAddr;
    fn set_pc(&mut self, pc: LC3MemAddr);
    fn reg(&self, addr: RegAddr) -> LC3Word;
    fn set_reg(&mut self, addr: RegAddr, value: LC3Word);
    fn mem(&self, addr: LC3MemAddr) -> LC3Word;
    fn set_mem(&mut self, addr: LC3MemAddr, value: LC3Word);
    fn priority(&self) -> u8;
    fn set_priority(&mut self, priority: u8);
    fn privileged(&self) -> bool;
    fn set_privileged(&mut self, priviledged: bool);
    fn positive_cond(&self) -> bool;
    fn zero_cond(&self) -> bool;
    fn negative_cond(&self) -> bool;
    fn flag_positive(&mut self);
    fn flag_zero(&mut self);
    fn flag_negative(&mut self);
    fn clear_flags(&mut self);
    fn iter(&self) -> Self::FullIter<'_>;
    fn sparse_iter(&self) -> Self::SparseIter<'_>;
    fn halt(&mut self);
    fn unhalt(&mut self);
    fn is_halted(&self) -> bool;
    fn step(&mut self) -> Result<(), StepFailure>;
    fn populate<I: IntoIterator<Item = LC3Word>>(
        &mut self,
        start: LC3MemAddr,
        words: I,
    );

    // Provided methods
    fn processor_status_reg(&self) -> LC3Word { ... }
    fn set_processor_status_reg(&mut self, status_reg: LC3Word) { ... }
    fn cur_inst(&self) -> Option<InstructionEnum> { ... }
    fn interrupt(&mut self, vector: LC3Word, set_priority: Option<u8>) { ... }

}

Full LC3 simulator.

Required Associated Types

type FullIter<'a>: Iterator<Item = LC3Word> where Self: 'a

Produces all words in order from 0x0000.

type SparseIter<'a>: Iterator<Item = LC3MemLoc> where Self: 'a

Produces all words in order from 0x0000, possibly skipping 0x0000 words.

Required Methods

fn pc(&self) -> LC3MemAddr

Current program counter.

fn set_pc(&mut self, pc: LC3MemAddr)

Replace the current program counter.

fn reg(&self, addr: RegAddr) -> LC3Word

fn set_reg(&mut self, addr: RegAddr, value: LC3Word)

fn mem(&self, addr: LC3MemAddr) -> LC3Word

fn set_mem(&mut self, addr: LC3MemAddr, value: LC3Word)

fn priority(&self) -> u8

Current priority in [0, 7].

0 is the lowest priority, 7 is the highest.

fn set_priority(&mut self, priority: u8)

Sets priority if in [0, 7].

fn privileged(&self) -> bool

True if in supervisor mode, false if in user mode.

fn set_privileged(&mut self, priviledged: bool)

Sets to supervisor mode if true; to user mode if false.

fn positive_cond(&self) -> bool

Returns true if the positive flag is set.

fn zero_cond(&self) -> bool

Returns true if the zero flag is set.

fn negative_cond(&self) -> bool

Returns true if the negative flag is set.

fn flag_positive(&mut self)

Sets the positive flag.

fn flag_zero(&mut self)

Sets the zero flag.

fn flag_negative(&mut self)

Sets the negative flag.

fn clear_flags(&mut self)

Clears the sign flags.

fn iter(&self) -> Self::FullIter<'_>

fn sparse_iter(&self) -> Self::SparseIter<'_>

fn halt(&mut self)

Set the machine to a halted state.

fn unhalt(&mut self)

Unset the machine from a halted state, if set.

fn is_halted(&self) -> bool

Check if the machine is set to halted.

fn step(&mut self) -> Result<(), StepFailure>

Processes the instruction at Self::pc.

fn populate<I: IntoIterator<Item = LC3Word>>( &mut self, start: LC3MemAddr, words: I, )

Fill the lines from start with words.

Provided Methods

fn processor_status_reg(&self) -> LC3Word

Returns the current processor status register value.

fn set_processor_status_reg(&mut self, status_reg: LC3Word)

Restores the processor status register value.

fn cur_inst(&self) -> Option<InstructionEnum>

Return the instruction at Self::pc, if any.

fn interrupt(&mut self, vector: LC3Word, set_priority: Option<u8>)

Initiates the interrupt service routine for vector.

set_priority is Some on I/O device interrupts, None on exceptions.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.