lc3sim_project

Module util

Functions

apply_offset

format_all_word_bits

Format all bits from idx byte of a LC3 word into a nice to read format.

format_bits

Format bits from a single byte into a nice to read format.

format_word_bits

Format bits from idx byte of a LC3 word into a nice to read format.

shift_to_signed

shift_to_unsigned